requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
SQLAlchemy==2.0.23
cryptography==41.0.7
Werkzeug==3.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9